refactor: reorganize scripts into vipat_cli_scripts directory and update paths to avoid generic naming and potential import conflicts, write poetry.lock
